Detect kernel variant preference.
KernelPicker detects the kernel variant preference from two sources,
with a fallback to the default:

  1. /etc/dnf/plugins/kernelpicker.conf
  2. uname -r
  3. default (6.1)

The format of the configuration file is as follows:

  $ cat /etc/dnf/plugins/kernelpicker.conf
  [main]
  variant=

KernelPicker.VARIANTS is the available values that can be used in the
config file, and the corresponding major version is set at the sack()
method, where all available packages are loaded.

The reason why we set the major version in sack() is to support the
'latest' value in the next patch and to load the latest available
kernel version in case we release a new LTS kernel, for example.

We include this empty configuration as an example.  This is an invalid
configuration and ignored by KernelPicker, so it’s fine.

  $ dnf info kernel -v
  ...
  Ignoring kernel variant in kernelpicker.conf: ''
  ...
  Kernel variant: 6.1

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>