Filter kernel packages.
We need to prevent LTS kernel users from unintentionally upgrading to
a different major version.

The .get_excluded_base() method applies the conditions generated by
self.get_filter_query() to the packages in PACKAGE_NAMES, filtering
them according to the configuration.

For example, if 6.1 is preferred, it will exclude packages that are
below 6.1.0 or above 6.2.0.

Test:

on AL2023 instance with 6.1 kernel and kernel-headers,

  $ uname -r
  6.1.132-147.221.amzn2023.x86_64

  $ rpm -q kernel kernel-headers
  kernel-6.1.131-143.221.amzn2023.x86_64
  kernel-headers-6.1.131-143.221.amzn2023.x86_64

without KernelPicker, updating kernel-headers will select the 6.12 version,

  $ sudo dnf update kernel-headers
  ...
  Upgrading:
   kernel-headers     x86_64     6.12.22-27.96.amzn2023     amazonlinux     1.5 M

with KernelPicker, the preferred version will be selected.

  $ sudo dnf update kernel-headers -v
  ...
  Filtered packages
  ...
    kernel-headers-6.12.20.23.97.amzn2023.x86_64 (amazonlinux)
    kernel-headers-6.12.22.27.96.amzn2023.x86_64 (amazonlinux)
  ...
  Upgrading:
   kernel-headers    x86_64    6.1.132-147.221.amzn2023      amazonlinux    1.4 M

Link: https://dnf.readthedocs.io/en/latest/api_plugins.html#dnf.Plugin.sack
Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>