Filter livepatch.
The naming convention for livepatch packages is special, as it includes
the kernel version in the package Name, not in Version.

This is to avoid installing a livepatch version that does not match the
running kernel.

The .get_excluded_livepatch() method excludes livepatches with a major
version different from the configuration.

While it is possible to restrict to the version that matches the running
kernel based on $(uname -r), this approach is not recommended.

For example, if a user wants to install a new kernel and the corresponding
livepatch simultaneously and reboot, the new livepatch would be excluded
from the search results, meaning it would need to be installed again after
the reboot.

Without this change:

  $ dnf list *livepatch* | awk '{print $1}'
  Installed
  kernel-livepatch-repo-s3.noarch
  Available
  kernel-livepatch-6.1.10-15.42.x86_64
  ...
  kernel-livepatch-6.1.97-104.177.x86_64
  kernel-livepatch-6.12.20-23.97.x86_64
  kernel-livepatch-6.12.22-27.96.x86_64
  kernel-livepatch-repo-cdn.noarch

With this change:

When 6.1 is configured, only 6.12 livepatches do not hit.

  $ sudo dnf kernelpicker 6.1
  variant: 6.1

  $ dnf list *livepatch* | awk '{print $1}'
  Installed
  kernel-livepatch-repo-s3.noarch
  Available
  kernel-livepatch-6.1.10-15.42.x86_64
  ...
  kernel-livepatch-6.1.97-104.177.x86_64
  kernel-livepatch-repo-cdn.noarch

When 6.12 is configured, only 6.12 livepatches hit.

  $ sudo dnf kernelpicker 6.12
  variant: 6.12

  $ dnf list *livepatch* | awk '{print $1}'
  Installed
  kernel-livepatch-repo-s3.noarch
  Available
  kernel-livepatch-6.12.20-23.97.x86_64
  kernel-livepatch-6.12.22-27.96.x86_64
  kernel-livepatch-repo-cdn.noarch

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>