Filter namespaced kernel packages.
For AL2023, several packages related to the 6.12 kernel are namespaced.

For example, the 6.12 kernel is provided as kernel6.12 (without a dash,
as kernel6.12.6.12.XXX.YYY rather than kernel-6.12.XX.YYY).

This is to prevent unintended kernel upgrades in AL2023, which is
composed of a single repository.

Let's prevent 6.1 kernel users from upgrading to the 6.12 kernel unless
configured.

Note that 6.12 kernel users will not be affected by this change, even
without configuration, because the .config() method detects the preference
from $(uname -r).

Test:

on 6.1 kernel

  $ uname -r
  6.1.132-147.221.amzn2023.x86_64

without this change:

  $ sudo dnf install kernel6.12 -v
  ...
  Installing:
   kernel6.12      x86_64      6.12.22-27.96.amzn2023      amazonlinux       35 M

with this change:

  $ sudo dnf install kernel6.12 -v
  ...
  Filtered packages
    kernel6.12-6.12.20.23.97.amzn2023.x86_64 (amazonlinux)
    kernel6.12-6.12.22.27.96.amzn2023.x86_64 (amazonlinux)
  ...
  All matches were filtered out by exclude filtering for argument: kernel6.12
  Error: Unable to find a match: kernel6.12

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>