Don't exclude installed packages for dnf upgrade.
While testing the upgrade from 6.1 to 6.12 using non-namespaced
packages, I noticed that dnf install worked, but dnf upgrade did not.

  $ rpm -q kernel-headers
  kernel-headers-6.1.129-138.220.amzn2023.x86_64

  $ sudo dnf kernelpicker 6.12
  variant: 6.12

  $ sudo dnf upgrade kernel-headers -v
  Kernel variant: 6.12
  ...
  Filtered packages
    kernel-6.1.129.138.220.amzn2023.x86_64 (@System)
    kernel-headers-6.1.129.138.220.amzn2023.x86_64 (@System)  <---
  ...
  No match for argument: kernel-headers
  Package kernel-headers available, but not installed.
  Error: No packages marked for upgrade.

As indicated by the message, it seemed like dnf was under the illusion
that the kernel-headers package was not installed.

This was caused by the fact that the currently installed kernel-headers
package was excluded from the query.

To avoid this issue, let's ensure that the installed packages are not
excluded from the query.

Fixes: d50236375018 ("Filter kernel packages.")
Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>