Only include the running kernel version for dnf update.
Let’s say we release a new non-namespaced 6.12 kernel package.

6.1 users would be unintentionally forced to install the 6.12 kernel.

We can prevent the actual upgrade by leaving the grub entries unchanged.
After rebooting, the 6.1 kernel would boot.

However, the next dnf update would detect the 6.12 kernel as installed,
even though the running kernel is still 6.1, and we only want to apply
updates for 6.1.

If we didn’t filter out the not-running-but-installed 6.12 kernel,
the plugin would allow users to upgrade to the 6.12 kernel.

For such a case, only include the running kernel package in the dnf
update.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>