Add .resolved() template.
Some kernel 6.12 subpackages are namespaced, but others are not.

This triggered unwanted behaviour when we assigned CVEs to kernel6.12.

Then, its non-namespaced subpackages are also marked with the CVE, and
dnf upgrade --security will update the subpackages to 6.12 one even when
the running kernel is 6.1.

We want to fix up this situation by adding the corresponding subpackages
to the DNF transaction, which can be done in self.resolved() by calling
self.base.resolve().

Note that self.base.resolve() will call self.resolved() again, so we
need a mechanism to stop infinite recursion.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>