Fix up mixed non-namespaced subpackages.
If a subpackage is installed but its version is different from the
installing/installed kernel, self.install_non_namespaced_packages()
tries to install the matching version of the subpackage.

It can be done by specifying the wanted package's NVR to
self.base.install().

Note that we set strict=False to avoid hard failure when its dependency
resolution fails.

In this example, kernel is upgrading to the latest and kernel-headers
is downgraded to the corresponding version.

  $ rpm -qa | grep kernel
  kernel-6.1.129-138.220.amzn2023.x86_64
  kernel-headers-6.12.22-27.96.amzn2023.x86_64

  $ sudo make && sudo dnf kernelpicker 6.1 && sudo dnf update -v
  ...
  Installing:
   kernel            x86_64    6.1.132-147.221.amzn2023     amazonlinux      31 M
  Downgrading:
   kernel-headers    x86_64    6.1.132-147.221.amzn2023     amazonlinux     1.4 M

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>